// src/app/entities/components/pages/threats/entities/components/threat-info/entities/threat-info-field.enum.ts
/**
 * Поля на странице информации об угрозе
 * 
 * @property TITLE - заголовок
 * @property DESCRIPTION - описание
 * @property CONSEQUANCES - последствия
 * @property PROTECTION - защита от угрозы
 * @property LINKS - полезные ссылки
 */
export enum EThreatInfoField {
    TITLE = 'title',
    DESCRIPTION = 'description',
    CONSEQUANCES = 'consequances',
    PROTECTION = 'protection',
    LINKS = 'links'
}